services/ingest/requirements.txt
fastapi
gunicorn
uvicorn[standard]
google-cloud-storage
google-cloud-bigquery
google-cloud-pubsub
google-cloud-videointelligence
opencv-python-headless
Pillow
requests

